Refactor: reorganize folder structure
