+++ src/components/multiplechoice/Cheers.tsx
