+++ src/components/scores/Result.tsx
