+++ src/components/scores/StarRating.test.tsx
