+++ src/components/scores/StarRating.tsx
